starship: remove aws session duration
This is just noise to me for 99% of the day. Would be nice to have it
ONLY when it's expired or about to expire, but seems that's not
currently supported.

I've had this in my config forever but it wasn't showing because it
wasn't compatible with AWS SSO, but support for this was recently
merged.